requirements.txt
Flask>=2.3.0
Flask-Limiter>=3.5.0
python-dotenv>=1.0.0
waitress>=2.1.2
redis